p31-surrogate-backend/src/nexus/register_pipeline.py
#!/usr/bin/env python3
"""
P31 Surrogate Backend - 3-Register Nexus Machine Pipeline
=======================================================

Implements the 3-Register Nexus Machine for deterministic execution:
- Register P (Past): Immutable Rulesets via Vertex AI Context Caching
- Register N (Now): Real-time telemetry and Node One biometric data
- Register U (Universe): Coupling logic where Register P constrains Register N

Author: P31 Labs
License: MIT
"""

import os
import time
import json
import logging
from typing import Dict, Any, Optional, Tuple
from google.cloud import aiplatform
from litellm import completion

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class NexusMachine:
    """3-Register Nexus Machine for deterministic cognitive processing"""
    
    def __init__(self):
        # Register P (Past): Immutable Rulesets via Vertex AI Context Caching
        self.register_p_cache_id = None
        self.register_p_initialized = False
        
        # Register N (Now): Real-time state
        self.register_n_state = {}
        
        # Register U (Universe): Coupling logic
        self.register_u_active = False
        
        # Initialize Vertex AI Context Caching
        self._initialize_vertex_ai()
        
        logger.info("🔗 Nexus Machine Initialized")
        logger.info(f"📚 Register P Cache ID: {self.register_p_cache_id}")
        logger.info(f"⏱️  Register N State: Active")
        logger.info(f"🌐 Register U Coupling: Ready")

    def _initialize_vertex_ai(self) -> None:
        """Initialize Vertex AI Context Caching for Register P"""
        try:
            # Configure Vertex AI
            project_id = os.getenv("VERTEX_AI_PROJECT", "p31-surrogate")
            location = os.getenv("VERTEX_AI_LOCATION", "us-central1")
            
            # Initialize Vertex AI
            aiplatform.init(project=project_id, location=location)
            
            # Mocking Vertex AI Context Caching initialization
            # In production, this would upload CognitivePassport-v2_5.md to GCP
            # and return a cache_id for immutable ruleset storage
            self.register_p_cache_id = "cache-v2_5-immutable"
            self.register_p_initialized = True
            
            logger.info(f"📚 Vertex AI Context Caching Initialized")
            logger.info(f"📍 Project: {project_id}")
            logger.info(f"🌍 Location: {location}")
            logger.info(f"🆔 Cache ID: {self.register_p_cache_id}")
            
        except Exception as e:
            logger.error(f"❌ Vertex AI initialization failed: {e}")
            # Fallback to local ruleset storage
            self.register_p_cache_id = "local-cache-fallback"
            self.register_p_initialized = False

    def ingest_register_n(self, prompt: str, telemetry: Dict[str, Any]) -> Dict[str, Any]:
        """
        Register N (Now): Ingest real-time conversational inputs and Node One biometric data
        
        Args:
            prompt (str): User's immediate intent
            telemetry (Dict[str, Any]): Biometric and contextual telemetry
            
        Returns:
            Dict[str, Any]: Register N state
        """
        try:
            register_n_state = {
                "immediate_intent": prompt,
                "biometric_telemetry": telemetry,
                "timestamp": time.time(),
                "session_id": f"session_{int(time.time())}",
                "context_window": self._build_context_window(telemetry)
            }
            
            self.register_n_state = register_n_state
            
            logger.info(f"⏱️  Register N Ingested:")
            logger.info(f"  🎯 Intent: {prompt[:50]}...")
            logger.info(f"  📊 Telemetry Keys: {list(telemetry.keys())}")
            logger.info(f"  ⏰ Timestamp: {register_n_state['timestamp']}")
            
            return register_n_state
            
        except Exception as e:
            logger.error(f"❌ Register N ingestion failed: {e}")
            return {"error": str(e)}

    def _build_context_window(self, telemetry: Dict[str, Any]) -> Dict[str, Any]:
        """Build context window from telemetry data"""
        context_window = {
            "voltage_level": telemetry.get("voltage_level", 50),
            "entropy_hash": telemetry.get("entropy_hash", "unknown"),
            "biometric_confidence": telemetry.get("confidence", 0.8),
            "environmental_context": telemetry.get("environment", {}),
            "temporal_context": {
                "session_duration": telemetry.get("session_duration", 0),
                "last_interaction": telemetry.get("last_interaction", time.time())
            }
        }
        return context_window

    def execute_register_u(self, register_n_state: Dict[str, Any]) -> str:
        """
        Register U (Universe): Coupling logic where Register P constrains Register N
        
        Args:
            register_n_state (Dict[str, Any]): Current Register N state
            
        Returns:
            str: Deterministic output
        """
        try:
            # Build system constraint using Register P cache
            system_constraint = self._build_system_constraint(register_n_state)
            
            # Execute constrained completion
            response = completion(
                model="claude-3-5-sonnet",
                messages=[
                    {"role": "system", "content": system_constraint},
                    {"role": "user", "content": register_n_state["immediate_intent"]}
                ],
                temperature=0.1,  # Deterministic output
                max_tokens=500
            )
            
            deterministic_output = response.choices[0].message.content
            self.register_u_active = True
            
            logger.info(f"🌐 Register U Executed:")
            logger.info(f"  📚 Constraint: {system_constraint[:100]}...")
            logger.info(f"  🎯 Input: {register_n_state['immediate_intent'][:50]}...")
            logger.info(f"  ✅ Output: {deterministic_output[:100]}...")
            
            return deterministic_output
            
        except Exception as e:
            logger.error(f"❌ Register U execution failed: {e}")
            return f"ERROR: Register U execution failed - {str(e)}"

    def _build_system_constraint(self, register_n_state: Dict[str, Any]) -> str:
        """Build system constraint using Register P cache and Register N state"""
        
        # Base constraint from Register P (immutable rulesets)
        base_constraint = f"""
USE CACHE_ID: {self.register_p_cache_id}. ENFORCE RULES ON IMMEDIATE STATE.

CRITICAL CONSTRAINTS:
1. MAINTAIN ENGINEER-POET VOICE: Technical precision + human empathy
2. NO NAVAL/MILITARY METAPHORS: Use electrical/quantum/geometry metaphors only
3. TIGER STYLE EXECUTION: Aggressive, high-velocity, Execute-Don't-Narrate
4. VOLTAGE AWARENESS: Consider current voltage_level ({register_n_state['context_window']['voltage_level']}/100)
5. BIOMETRIC INTEGRITY: Respect confidence_score ({register_n_state['context_window']['biometric_confidence']})

IMMEDIATE STATE TO CONSTRAIN:
- Intent: {register_n_state['immediate_intent'][:100]}...
- Session: {register_n_state['session_id']}
- Environment: {register_n_state['context_window']['environmental_context']}

OUTPUT REQUIREMENTS:
- Actionable response only
- No explanations unless explicitly requested
- Maintain deterministic behavior
- Preserve cognitive coherence
"""
        
        return base_constraint.strip()

    def process_pipeline(self, prompt: str, telemetry: Dict[str, Any]) -> Dict[str, Any]:
        """
        Execute the complete 3-Register Nexus Machine pipeline
        
        Args:
            prompt (str): User's immediate intent
            telemetry (Dict[str, Any]): Biometric and contextual telemetry
            
        Returns:
            Dict[str, Any]: Complete pipeline result
        """
        try:
            logger.info("🔄 Executing 3-Register Nexus Machine Pipeline...")
            
            # Step 1: Register N (Now) - Ingest real-time data
            register_n_state = self.ingest_register_n(prompt, telemetry)
            
            # Step 2: Register U (Universe) - Execute constrained processing
            output = self.execute_register_u(register_n_state)
            
            # Step 3: Build complete result
            result = {
                "pipeline_execution": {
                    "timestamp": time.time(),
                    "register_p_cache_id": self.register_p_cache_id,
                    "register_n_session_id": register_n_state.get("session_id"),
                    "register_u_active": self.register_u_active
                },
                "input": {
                    "prompt": prompt,
                    "telemetry": telemetry
                },
                "output": {
                    "response": output,
                    "deterministic": True,
                    "constrained_by_ruleset": self.register_p_initialized
                },
                "metadata": {
                    "voltage_level": register_n_state['context_window']['voltage_level'],
                    "biometric_confidence": register_n_state['context_window']['biometric_confidence'],
                    "execution_time": time.time() - register_n_state['timestamp']
                }
            }
            
            logger.info("✅ 3-Register Nexus Machine Pipeline Complete")
            return result
            
        except Exception as e:
            logger.error(f"💥 Pipeline execution failed: {e}")
            return {
                "error": str(e),
                "pipeline_execution": {"failed": True}
            }

    def health_check(self) -> Dict[str, Any]:
        """Return health status of Nexus Machine"""
        return {
            "status": "healthy" if self.register_p_initialized else "degraded",
            "register_p": {
                "initialized": self.register_p_initialized,
                "cache_id": self.register_p_cache_id,
                "vertex_ai_project": os.getenv("VERTEX_AI_PROJECT", "not set")
            },
            "register_n": {
                "active_session": bool(self.register_n_state),
                "last_update": self.register_n_state.get("timestamp", "none")
            },
            "register_u": {
                "active": self.register_u_active,
                "last_execution": time.time() if self.register_u_active else None
            },
            "timestamp": time.time()
        }

def main():
    """Main entry point for testing Nexus Machine"""
    logger.info("🔗 Initializing P31 3-Register Nexus Machine...")
    
    try:
        nexus = NexusMachine()
        
        # Perform health check
        health = nexus.health_check()
        logger.info(f"🏥 Health Status: {health}")
        
        # Test pipeline with sample data
        test_prompt = "Generate a Python script for processing biometric data"
        test_telemetry = {
            "voltage_level": 65,
            "entropy_hash": "test_hash_123",
            "confidence": 0.9,
            "environment": {"device": "node_one", "location": "home"},
            "session_duration": 120,
            "last_interaction": time.time() - 30
        }
        
        logger.info("🧪 Testing pipeline execution...")
        result = nexus.process_pipeline(test_prompt, test_telemetry)
        
        logger.info(f"📊 Pipeline Result: {result['output']['deterministic']}")
        logger.info("✅ Nexus Machine Ready")
        
    except Exception as e:
        logger.error(f"💥 Nexus Machine failed to start: {e}")
        exit(1)

if __name__ == "__main__":
    main()