p31-surrogate-backend/requirements.txt
# P31 Surrogate Backend Dependencies

# Core FastAPI and web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2

# PGLite Database
@electric-sql/pglite==0.1.5
psycopg2-binary==2.9.9

# Redis client for atomic operations and Catcher's Mitt
redis==5.0.1
aioredis==2.0.1

# Semantic Routing and LLM Integration
litellm==1.49.0
openai==1.35.12
google-cloud-aiplatform==1.52.0

# Vector Database and RAG
chromadb==0.5.3
sentence-transformers==3.0.1
faiss-cpu==1.7.4

# Procedural Memory and RAFT/PEFT
transformers==4.44.2
peft==0.12.0
datasets==2.21.0

# HTTP client for testing and Node One integration
aiohttp==3.9.5
requests==2.32.3

# Background task processing
celery==5.4.0
redis-celery==5.4.0

# Configuration and environment
python-dotenv==1.0.1
pyyaml==6.0.1

# Logging and monitoring
structlog==24.2.0
prometheus-client==0.21.1

# Security and authentication
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.4.2
ruff==0.5.4

# Additional utilities
click==8.1.7
typer==0.12.5
rich==13.9.4